Extract key_encode, render, and selection modules from main.rs (#68)
Phase 1 of main.rs decomposition: extract three free-function groups
into dedicated modules with no behavior changes.

- key_encode.rs (96 lines): egui key → terminal byte encoding
- render.rs (333 lines): pane rendering, exit overlay, color conversion
- selection.rs (129 lines): pointer-to-cell mapping, word bounds, text extraction

main.rs: 4,644 → 4,114 lines (-530)
