+++ crates/amux-app/src/main.rs
@@ -1,6 +1,9 @@
 mod ipc_dispatch;
+mod key_encode;
 mod managed_pane;
 mod menu_bar;
+mod render;
+mod selection;
 mod sidebar;
 mod system_notify;
 mod theme;
@@ -20,15 +23,12 @@ use amux_notify::{
     flash_alpha, FlashReason, NotificationSource, NotificationStore, FLASH_DURATION,
 };
 use amux_session::SessionData;
-use amux_term::color::resolve_color;
 use amux_term::config::AmuxTermConfig;
 use amux_term::font;
 use amux_term::osc::NotificationEvent;
 use amux_term::pane::TerminalPane;
 use managed_pane::*;
 use portable_pty::CommandBuilder;
-use wezterm_surface::{CursorShape, CursorVisibility};
-use wezterm_term::color::SrgbaTuple;
 
 #[cfg(feature = "gpu-renderer")]
 use amux_render_gpu::GpuRenderer;
@@ -1865,7 +1865,7 @@ impl AmuxApp {
             .or(managed.selection.as_ref())
             .cloned();
         let surface = managed.active_surface_mut();
-        render_pane(
+        render::render_pane(
             ui,
             &mut surface.pane,
             content_rect,
@@ -1883,7 +1883,7 @@ impl AmuxApp {
 
         // Render exit overlay when process has exited
         if let Some(exit_info) = &surface.exited {
-            render_exit_overlay(ui, content_rect, &exit_info.message, self.font_size);
+            render::render_exit_overlay(ui, content_rect, &exit_info.message, self.font_size);
         }
 
         // Render copy mode cursor overlay
@@ -3638,7 +3638,7 @@ impl AmuxApp {
                                                 .color(egui::Color32::from_gray(140)),
                                         );
                                     }
-                                    let age = format_duration(notif.created_at.elapsed());
+                                    let age = render::format_duration(notif.created_at.elapsed());
                                     ui.label(
                                         egui::RichText::new(age)
                                             .small()
@@ -3768,7 +3768,8 @@ impl AmuxApp {
 
         let (cols, _) = managed.active_surface().pane.dimensions();
         let (start, end) = sel.normalized();
-        let text = extract_selection_text(&managed.active_surface().pane, start, end, cols);
+        let text =
+            selection::extract_selection_text(&managed.active_surface().pane, start, end, cols);
 
         if text.is_empty() {
             return false;
@@ -3870,7 +3871,7 @@ impl AmuxApp {
                     return true;
                 }
 
-                let (col, stable_row) = pointer_to_cell(
+                let (col, stable_row) = selection::pointer_to_cell(
                     pos,
                     content_rect,
                     cell_width,
@@ -3896,9 +3897,12 @@ impl AmuxApp {
                 let (anchor, end, mode) = match self.click_count {
                     2 => {
                         // Word selection
-                        let text =
-                            line_text_string(&managed.active_surface().pane, stable_row, cols);
-                        let (wstart, wend) = word_bounds_in_line(&text, col);
+                        let text = selection::line_text_string(
+                            &managed.active_surface().pane,
+                            stable_row,
+                            cols,
+                        );
+                        let (wstart, wend) = selection::word_bounds_in_line(&text, col);
                         (
                             (wstart, stable_row),
                             (wend, stable_row),
@@ -3939,7 +3943,7 @@ impl AmuxApp {
 
             if has_active_selection {
                 if let Some(pos) = pointer_pos {
-                    let (col, stable_row) = pointer_to_cell(
+                    let (col, stable_row) = selection::pointer_to_cell(
                         pos,
                         content_rect,
                         cell_width,
@@ -3957,19 +3961,20 @@ impl AmuxApp {
                                     sel.end = (col, stable_row);
                                 }
                                 SelectionMode::Word => {
-                                    let text = line_text_string(
+                                    let text = selection::line_text_string(
                                         &m.surfaces[m.active_surface_idx].pane,
                                         stable_row,
                                         cols,
                                     );
-                                    let (_, wend) = word_bounds_in_line(&text, col);
+                                    let (_, wend) = selection::word_bounds_in_line(&text, col);
                                     // Extend: keep anchor word start, update end word boundary
                                     if stable_row > sel.anchor.1
                                         || (stable_row == sel.anchor.1 && col >= sel.anchor.0)
                                     {
                                         sel.end = (wend, stable_row);
                                     } else {
-                                        let (wstart, _) = word_bounds_in_line(&text, col);
+                                        let (wstart, _) =
+                                            selection::word_bounds_in_line(&text, col);
                                         sel.end = (wstart, stable_row);
                                     }
                                 }
@@ -4068,7 +4073,7 @@ impl AmuxApp {
                     modifiers,
                     ..
                 } => {
-                    if let Some(bytes) = encode_egui_key(key, modifiers) {
+                    if let Some(bytes) = key_encode::encode_egui_key(key, modifiers) {
                         surface.scroll_offset = 0;
                         surface.scroll_accum = 0.0;
                         let _ = surface.pane.write_bytes(&bytes);
@@ -4106,539 +4111,4 @@ impl AmuxApp {
         }
         has_input
     }
-}
-
-// --- Selection helpers ---
-
-/// Convert a pointer position to terminal cell coordinates (col, stable_row).
-fn pointer_to_cell(
-    pointer_pos: egui::Pos2,
-    content_rect: egui::Rect,
-    cell_width: f32,
-    cell_height: f32,
-    scroll_offset: usize,
-    total_rows: usize,
-    visible_rows: usize,
-) -> (usize, usize) {
-    let col = ((pointer_pos.x - content_rect.min.x) / cell_width)
-        .floor()
-        .max(0.0) as usize;
-    let visible_row = ((pointer_pos.y - content_rect.min.y) / cell_height)
-        .floor()
-        .max(0.0) as usize;
-    let visible_row = visible_row.min(visible_rows.saturating_sub(1));
-    let stable_row = total_rows
-        .saturating_sub(visible_rows)
-        .saturating_sub(scroll_offset)
-        + visible_row;
-    (col, stable_row)
-}
-
-/// Find word boundaries around a column in a line's text.
-/// Returns (start_col, end_col) inclusive.
-fn word_bounds_in_line(line_text: &str, col: usize) -> (usize, usize) {
-    let chars: Vec<char> = line_text.chars().collect();
-    if chars.is_empty() || col >= chars.len() {
-        return (col, col);
-    }
-
-    let is_delim = |ch: char| WORD_DELIMITERS.contains(ch);
-    let at_delim = is_delim(chars[col]);
-
-    // Walk left
-    let mut start = col;
-    while start > 0 && is_delim(chars[start - 1]) == at_delim {
-        start -= 1;
-    }
-
-    // Walk right
-    let mut end = col;
-    while end + 1 < chars.len() && is_delim(chars[end + 1]) == at_delim {
-        end += 1;
-    }
-
-    (start, end)
-}
-
-/// Extract text from the terminal screen within a selection range.
-fn extract_selection_text(
-    pane: &TerminalPane,
-    start: (usize, usize),
-    end: (usize, usize),
-    cols: usize,
-) -> String {
-    let screen = pane.screen();
-    let lines = screen.lines_in_phys_range(start.1..end.1 + 1);
-    let mut result = String::new();
-
-    for (i, line) in lines.iter().enumerate() {
-        let row = start.1 + i;
-        let mut line_text = String::new();
-        for cell in line.visible_cells() {
-            let ci = cell.cell_index();
-            if ci >= cols {
-                break;
-            }
-            // Determine if this cell is in the selection
-            let in_sel = if start.1 == end.1 {
-                ci >= start.0 && ci <= end.0
-            } else if row == start.1 {
-                ci >= start.0
-            } else if row == end.1 {
-                ci <= end.0
-            } else {
-                true
-            };
-            if in_sel {
-                line_text.push_str(cell.str());
-            }
-        }
-
-        if i > 0 {
-            // Check if previous line was wrapped — if so, don't add newline
-            if i > 0 {
-                let prev_line = &lines[i - 1];
-                if !prev_line.last_cell_was_wrapped() {
-                    result.push('\n');
-                }
-            }
-        }
-        result.push_str(line_text.trim_end());
-    }
-
-    result
-}
-
-/// Build a flat string of a line's cell text for word boundary detection.
-fn line_text_string(pane: &TerminalPane, stable_row: usize, cols: usize) -> String {
-    let screen = pane.screen();
-    let lines = screen.lines_in_phys_range(stable_row..stable_row + 1);
-    if lines.is_empty() {
-        return String::new();
-    }
-    let line = &lines[0];
-    let mut text = String::new();
-    for cell in line.visible_cells() {
-        if cell.cell_index() >= cols {
-            break;
-        }
-        let s = cell.str();
-        if s.is_empty() {
-            text.push(' ');
-        } else {
-            text.push_str(s);
-        }
-    }
-    text
-}
-
-// --- Rendering ---
-
-#[allow(clippy::too_many_arguments)]
-fn render_pane(
-    ui: &mut egui::Ui,
-    pane: &mut TerminalPane,
-    rect: egui::Rect,
-    is_focused: bool,
-    scroll_offset: usize,
-    selection: Option<&SelectionState>,
-    font_size: f32,
-    find_highlights: &[(usize, usize, usize)],
-    current_highlight: Option<usize>,
-    #[cfg(feature = "gpu-renderer")] gpu_renderer: Option<&GpuRenderer>,
-    #[cfg(feature = "gpu-renderer")] pane_id: u64,
-) {
-    // GPU renderer path: build snapshot, emit a paint callback, and return early.
-    #[cfg(feature = "gpu-renderer")]
-    if let Some(gpu) = gpu_renderer {
-        let (actual_cols, actual_rows) = pane.dimensions();
-        if actual_cols == 0 || actual_rows == 0 {
-            return;
-        }
-        let palette = pane.palette();
-        let cursor = pane.cursor();
-        let screen = pane.screen();
-        let gpu_selection = selection.map(|sel| {
-            let (start, end) = sel.normalized();
-            amux_render_gpu::snapshot::SelectionRange { start, end }
-        });
-        let seqno = pane.current_seqno();
-        let snapshot = amux_render_gpu::TerminalSnapshot::from_screen(
-            screen,
-            &palette,
-            &cursor,
-            actual_cols,
-            actual_rows,
-            scroll_offset,
-            is_focused,
-            gpu_selection,
-            pane_id,
-            seqno,
-            find_highlights.to_vec(),
-            current_highlight,
-        );
-        let pixels_per_point = ui.ctx().pixels_per_point();
-        let callback = gpu.paint_callback(rect, snapshot, pixels_per_point);
-        ui.painter().add(egui::Shape::Callback(callback));
-        return;
-    }
-
-    let font_id = egui::FontId::monospace(font_size);
-    let cell_width = ui.fonts(|f| f.glyph_width(&font_id, 'M'));
-    let cell_height = ui.fonts(|f| f.row_height(&font_id));
-
-    let (actual_cols, actual_rows) = pane.dimensions();
-    if actual_cols == 0 || actual_rows == 0 {
-        return;
-    }
-
-    let palette = pane.palette();
-    let cursor = pane.cursor();
-    let screen = pane.screen();
-
-    let painter = ui.painter();
-    let origin = rect.min;
-    let bg_default = srgba_to_egui(palette.background);
-
-    // Fill the full allocated rect first to avoid artifacts when terminal is smaller
-    painter.rect_filled(rect, 0.0, bg_default);
-
-    // Dim unfocused panes with a semi-transparent overlay
-    if !is_focused {
-        let dim_overlay = egui::Color32::from_rgba_unmultiplied(0, 0, 0, 100);
-        painter.rect_filled(rect, 0.0, dim_overlay);
-    }
-
-    let total = screen.scrollback_rows();
-    let end = total.saturating_sub(scroll_offset);
-    let start = end.saturating_sub(actual_rows);
-    let lines = screen.lines_in_phys_range(start..end);
-
-    for (row_idx, line) in lines.iter().enumerate() {
-        let y = origin.y + row_idx as f32 * cell_height;
-        if y + cell_height < rect.min.y || y > rect.max.y {
-            continue;
-        }
-
-        for cell_ref in line.visible_cells() {
-            let col_idx = cell_ref.cell_index();
-            if col_idx >= actual_cols {
-                break;
-            }
-
-            let x = origin.x + col_idx as f32 * cell_width;
-            if x + cell_width < rect.min.x || x > rect.max.x {
-                continue;
-            }
-
-            let attrs = cell_ref.attrs();
-            let reverse = attrs.reverse();
-            let mut bg =
-                srgba_to_egui(resolve_color(&attrs.background(), &palette, false, reverse));
-            let mut fg = srgba_to_egui(resolve_color(&attrs.foreground(), &palette, true, reverse));
-
-            // Selection: swap fg/bg for selected cells (reverse video)
-            let stable_row = start + row_idx;
-            if let Some(sel) = selection {
-                if sel.contains(col_idx, stable_row) {
-                    std::mem::swap(&mut fg, &mut bg);
-                    // Ensure selected empty cells have visible bg
-                    if bg == bg_default {
-                        bg = srgba_to_egui(palette.foreground);
-                        fg = bg_default;
-                    }
-                }
-            }
-
-            // Find/search highlighting
-            for (i, &(h_row, h_start, h_end)) in find_highlights.iter().enumerate() {
-                if h_row == stable_row && col_idx >= h_start && col_idx < h_end {
-                    if current_highlight == Some(i) {
-                        bg = egui::Color32::from_rgb(255, 153, 0); // orange
-                    } else {
-                        bg = egui::Color32::from_rgba_unmultiplied(255, 255, 0, 180);
-                        // yellow
-                    }
-                    fg = egui::Color32::BLACK;
-                    break;
-                }
-            }
-
-            if bg != bg_default {
-                painter.rect_filled(
-                    egui::Rect::from_min_size(
-                        egui::pos2(x, y),
-                        egui::vec2(cell_width, cell_height),
-                    ),
-                    0.0,
-                    bg,
-                );
-            }
-
-            let text = cell_ref.str();
-            if !text.is_empty() && text != " " {
-                painter.text(
-                    egui::pos2(x, y),
-                    egui::Align2::LEFT_TOP,
-                    text,
-                    font_id.clone(),
-                    fg,
-                );
-            }
-        }
-    }
-
-    // Draw cursor
-    if is_focused
-        && scroll_offset == 0
-        && cursor.visibility == CursorVisibility::Visible
-        && cursor.y >= 0
-        && (cursor.y as usize) < actual_rows
-        && cursor.x < actual_cols
-    {
-        let cx = origin.x + cursor.x as f32 * cell_width;
-        let cy = origin.y + cursor.y as f32 * cell_height;
-        let cursor_color = srgba_to_egui(palette.cursor_bg);
-
-        match cursor.shape {
-            CursorShape::BlinkingBar | CursorShape::SteadyBar => {
-                let bar_rect =
-                    egui::Rect::from_min_size(egui::pos2(cx, cy), egui::vec2(2.0, cell_height));
-                painter.rect_filled(bar_rect, 0.0, cursor_color);
-            }
-            CursorShape::BlinkingUnderline | CursorShape::SteadyUnderline => {
-                let underline_rect = egui::Rect::from_min_size(
-                    egui::pos2(cx, cy + cell_height - 2.0),
-                    egui::vec2(cell_width, 2.0),
-                );
-                painter.rect_filled(underline_rect, 0.0, cursor_color);
-            }
-            CursorShape::Default | CursorShape::BlinkingBlock | CursorShape::SteadyBlock => {
-                let cursor_rect = egui::Rect::from_min_size(
-                    egui::pos2(cx, cy),
-                    egui::vec2(cell_width, cell_height),
-                );
-                let cursor_fg = srgba_to_egui(palette.cursor_fg);
-                painter.rect_filled(cursor_rect, 0.0, cursor_color);
-
-                let cursor_line_idx = cursor.y as usize;
-                if cursor_line_idx < lines.len() {
-                    let line = &lines[cursor_line_idx];
-                    for cell_ref in line.visible_cells() {
-                        if cell_ref.cell_index() == cursor.x {
-                            let text = cell_ref.str();
-                            if !text.is_empty() && text != " " {
-                                painter.text(
-                                    egui::pos2(cx, cy),
-                                    egui::Align2::LEFT_TOP,
-                                    text,
-                                    font_id.clone(),
-                                    cursor_fg,
-                                );
-                            }
-                            break;
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    // Scroll indicator
-    if scroll_offset > 0 {
-        let indicator = format!("[+{}]", scroll_offset);
-        let indicator_font = egui::FontId::monospace(font_size * 0.8);
-        let text_color = egui::Color32::from_rgba_unmultiplied(255, 200, 50, 200);
-        let bg_color = egui::Color32::from_rgba_unmultiplied(40, 40, 40, 180);
-
-        let galley = painter.layout_no_wrap(indicator, indicator_font, text_color);
-        let text_size = galley.size();
-        let padding = 4.0;
-        let indicator_rect = egui::Rect::from_min_size(
-            egui::pos2(
-                rect.right() - text_size.x - padding * 2.0,
-                rect.top() + padding,
-            ),
-            egui::vec2(text_size.x + padding * 2.0, text_size.y + padding),
-        );
-        painter.rect_filled(indicator_rect, 3.0, bg_color);
-        painter.galley(
-            egui::pos2(
-                indicator_rect.left() + padding,
-                indicator_rect.top() + padding * 0.5,
-            ),
-            galley,
-            text_color,
-        );
-    }
-}
-
-/// Render a semi-transparent overlay at the bottom of a pane showing exit status
-/// and available actions (Enter to close, R to restart).
-fn render_exit_overlay(ui: &mut egui::Ui, rect: egui::Rect, message: &str, font_size: f32) {
-    let painter = ui.painter_at(rect);
-    let overlay_font = egui::FontId::monospace(font_size * 0.85);
-    let small_font = egui::FontId::monospace(font_size * 0.75);
-
-    let line1 = message;
-    let line2 = "Press Enter to close  |  R to restart";
-
-    let g1 = painter.layout_no_wrap(
-        line1.to_string(),
-        overlay_font.clone(),
-        egui::Color32::WHITE,
-    );
-    let g2 = painter.layout_no_wrap(line2.to_string(), small_font, egui::Color32::from_gray(160));
-
-    let text_width = g1.size().x.max(g2.size().x);
-    let text_height = g1.size().y + g2.size().y + 4.0;
-    let padding = 8.0;
-    let box_width = text_width + padding * 2.0;
-    let box_height = text_height + padding * 2.0;
-
-    let box_rect = egui::Rect::from_min_size(
-        egui::pos2(
-            rect.center().x - box_width / 2.0,
-            rect.bottom() - box_height - 12.0,
-        ),
-        egui::vec2(box_width, box_height),
-    );
-
-    // Semi-transparent background with border
-    painter.rect_filled(
-        box_rect,
-        4.0,
-        egui::Color32::from_rgba_unmultiplied(30, 30, 30, 220),
-    );
-    painter.rect_stroke(
-        box_rect,
-        4.0,
-        egui::Stroke::new(1.0, egui::Color32::from_gray(80)),
-        egui::StrokeKind::Outside,
-    );
-
-    // Center text lines
-    let x1 = box_rect.center().x - g1.size().x / 2.0;
-    let y1 = box_rect.min.y + padding;
-    painter.galley(egui::pos2(x1, y1), g1, egui::Color32::WHITE);
-
-    let x2 = box_rect.center().x - g2.size().x / 2.0;
-    let y2 = y1 + text_height - g2.size().y;
-    painter.galley(egui::pos2(x2, y2), g2, egui::Color32::from_gray(160));
-}
-
-fn srgba_to_egui(color: SrgbaTuple) -> egui::Color32 {
-    egui::Color32::from_rgba_unmultiplied(
-        (color.0 * 255.0).round() as u8,
-        (color.1 * 255.0).round() as u8,
-        (color.2 * 255.0).round() as u8,
-        (color.3 * 255.0).round() as u8,
-    )
-}
-
-fn format_duration(d: Duration) -> String {
-    let secs = d.as_secs();
-    if secs < 60 {
-        format!("{}s ago", secs)
-    } else if secs < 3600 {
-        format!("{}m ago", secs / 60)
-    } else {
-        format!("{}h ago", secs / 3600)
-    }
-}
-
-// --- Key encoding (egui events -> terminal bytes) ---
-
-fn encode_egui_key(key: &egui::Key, modifiers: &egui::Modifiers) -> Option<Vec<u8>> {
-    use amux_core::keys;
-
-    // Ctrl+key / Alt+key / Ctrl+Alt+key
-    if let Some(byte) = egui_ctrl_byte(key) {
-        if modifiers.ctrl && !modifiers.alt {
-            return Some(keys::encode_ctrl(byte));
-        }
-        if modifiers.alt && !modifiers.ctrl {
-            return Some(keys::encode_alt_char(byte - 1 + b'a'));
-        }
-        if modifiers.ctrl && modifiers.alt {
-            return Some(keys::encode_ctrl_alt(byte));
-        }
-    }
-
-    // Named keys — delegate to core encoder
-    let core_key = egui_key_to_core(key)?;
-    let mods = keys::Modifiers {
-        shift: modifiers.shift,
-        ctrl: modifiers.ctrl,
-        alt: modifiers.alt,
-    };
-    keys::encode_named(core_key, mods, false)
-}
-
-/// Map egui letter keys to their Ctrl control byte (A=0x01 .. Z=0x1a).
-fn egui_ctrl_byte(key: &egui::Key) -> Option<u8> {
-    match key {
-        egui::Key::A => Some(0x01),
-        egui::Key::B => Some(0x02),
-        egui::Key::C => Some(0x03),
-        egui::Key::D => Some(0x04),
-        egui::Key::E => Some(0x05),
-        egui::Key::F => Some(0x06),
-        egui::Key::G => Some(0x07),
-        egui::Key::H => Some(0x08),
-        egui::Key::I => Some(0x09),
-        egui::Key::J => Some(0x0a),
-        egui::Key::K => Some(0x0b),
-        egui::Key::L => Some(0x0c),
-        egui::Key::M => Some(0x0d),
-        egui::Key::N => Some(0x0e),
-        egui::Key::O => Some(0x0f),
-        egui::Key::P => Some(0x10),
-        egui::Key::Q => Some(0x11),
-        egui::Key::R => Some(0x12),
-        egui::Key::S => Some(0x13),
-        egui::Key::T => Some(0x14),
-        egui::Key::U => Some(0x15),
-        egui::Key::V => Some(0x16),
-        egui::Key::W => Some(0x17),
-        egui::Key::X => Some(0x18),
-        egui::Key::Y => Some(0x19),
-        egui::Key::Z => Some(0x1a),
-        _ => None,
-    }
-}
-
-/// Map egui Key to core NamedKey.
-fn egui_key_to_core(key: &egui::Key) -> Option<amux_core::keys::NamedKey> {
-    use amux_core::keys::NamedKey;
-    Some(match key {
-        egui::Key::Enter => NamedKey::Enter,
-        egui::Key::Tab => NamedKey::Tab,
-        egui::Key::Escape => NamedKey::Escape,
-        egui::Key::Backspace => NamedKey::Backspace,
-        egui::Key::Space => NamedKey::Space,
-        egui::Key::ArrowUp => NamedKey::ArrowUp,
-        egui::Key::ArrowDown => NamedKey::ArrowDown,
-        egui::Key::ArrowLeft => NamedKey::ArrowLeft,
-        egui::Key::ArrowRight => NamedKey::ArrowRight,
-        egui::Key::Home => NamedKey::Home,
-        egui::Key::End => NamedKey::End,
-        egui::Key::Insert => NamedKey::Insert,
-        egui::Key::Delete => NamedKey::Delete,
-        egui::Key::PageUp => NamedKey::PageUp,
-        egui::Key::PageDown => NamedKey::PageDown,
-        egui::Key::F1 => NamedKey::F1,
-        egui::Key::F2 => NamedKey::F2,
-        egui::Key::F3 => NamedKey::F3,
-        egui::Key::F4 => NamedKey::F4,
-        egui::Key::F5 => NamedKey::F5,
-        egui::Key::F6 => NamedKey::F6,
-        egui::Key::F7 => NamedKey::F7,
-        egui::Key::F8 => NamedKey::F8,
-        egui::Key::F9 => NamedKey::F9,
-        egui::Key::F10 => NamedKey::F10,
-        egui::Key::F11 => NamedKey::F11,
-        egui::Key::F12 => NamedKey::F12,
-        _ => return None,
-    })
 }
